refactor: folder structure modularized
